requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
langgraph==0.6.0
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.0
langgraph-sdk==0.2.0
langsmith==0.4.8
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
openai==1.97.1
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
psycopg2==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tenacity==9.1.2
tokenizers==0.21.4
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whisper==1.1.10
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
